+++ src/utils/formatDate.ts
@@ -1,0 +1,9 @@
+export function formatDate(iso: string): string {
+  const d = new Date(iso);
+  const Y = d.getFullYear();
+  const M = String(d.getMonth() + 1).padStart(2, '0');
+  const D = String(d.getDate()).padStart(2, '0');
+  const h = String(d.getHours()).padStart(2, '0');
+  const m = String(d.getMinutes()).padStart(2, '0');
+  return `${Y}.${M}.${D} ${h}:${m}`;
+}
